+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-09-16
+### Feat
+- Add support for localization using Unity Localization
+- Add support for multiple quest steps per QuestPoint
+
 ## [1.1.1] - 2024-09-13
 ### Feat
 - Add quick access methods to more easily get steps / quests from quests / questlines.